refactor: консолидация констант DISP_WIDTH и DISP_HEIGHT
- Удалено дублирование констант между io.rs и game/constants.rs
- Константы определены в game/constants.rs как вычисляемые
- Добавлен переэкспорт из io.rs для обратной совместимости
- Исправлены типы с u16 на usize для совместимости
- Обновлены тесты test_io.rs и test_cast_safety.rs

+++ src/menu/constants.rs
@@ -48,7 +48,7 @@ pub const MENU_RECORD_Y: u16 = 21;
 /// - Строки 8-15: Управление
 /// - Строки 17-20: Режимы игры
 /// - Строки 21-22: Рекорд
-pub const MENU: [&str; DISP_HEIGHT as usize] = [
+pub const MENU: [&str; DISP_HEIGHT] = [
     "                      ",
     "╔════════════════════╗",
     "║                    ║",
